created file for get_crypto_prices, moved helper method in there
